requirements.txt
bilibili-api-python
anthropic
obsws-python
python-dotenv
edge-tts
pygame
sounddevice
webrtcvad-wheels
numpy
httpx
pycaw
